This commit fixes a bug with OA3 and form param arrays in the body
Specifically, the following situation causes the bug:
- You are using OpenApi 3.x+
- You have a form parameter in the body of the request with type array
- you want to use the 'explode: true' option in the encoding block.

in legacy openApi, you would set collectionFormat: multi, which would then invoke our existing mustache templates, but now we need to use the explode option. In order to not re-write all of the mustache templates, but preserve the notion of explode, this commit adds a simple override of the isCollectionFormatMulti paramter that forces it to true if explode is set to true for this parameter.